userinfo, mute, unmute and ban, unban were tweaked a little

+++ main.py
@@ -18,7 +18,7 @@ def restart_bot():
     os.execv(sys.executable, ['python'] + sys.argv)
 
 
-@client.command(aliases=["reboot"])
+@client.command(aliases=["reboot", "reload"])
 @commands.has_permissions(administrator=True)
 async def restart(ctx):
     await ctx.reply("Success")
@@ -41,8 +41,6 @@ async def on_ready():
 
 @client.event
 async def on_message(message):
-    if message.content == 'w.hi':
-        await message.channel.send('Hello :)')
     await client.process_commands(message)
 
 
@@ -61,7 +59,7 @@ async def on_member_join(member):
     general = guild.get_channel(876085929117352016)
     verify = guild.get_channel(876102248185339925)
     await channel.send(
-        f'{member.mention} joined this server :O\nRemember, you can\'t leave')
+        f'{member.mention} joined this server :O\nRemember, you can\'t leave', delete_after=60)
     await member.send(
         f'Welcome to **{guild.name}**, {member.name}!\njust to make things clear, you\'ll die if you leave\nhttps://discord.gg/T5BZayunBB'
     )
@@ -86,6 +84,7 @@ snipe_message_id = None
 
 
 @client.event
+@commands.cooldown(1, 4, commands.BucketType.user)
 async def on_message_delete(message):
 
     global snipe_message_content
@@ -107,6 +106,7 @@ async def on_message_delete(message):
 
 
 @client.command()
+@commands.cooldown(1, 4, commands.BucketType.user)
 async def snipe(message):
     if snipe_message_content == None:
         await message.channel.send("Couldn't find anything to snipe!")
@@ -122,11 +122,13 @@ async def snipe(message):
 
 
 @client.command()
+@commands.cooldown(1, 4, commands.BucketType.user)
 async def ping(ctx):
     await ctx.send(f'{round(client.latency * 1000)}ms')
 
 
 @client.command(aliases=["av", "pfp"])
+@commands.cooldown(1, 4, commands.BucketType.user)
 async def avatar(ctx, member: discord.Member = None):
     if member == None:
         member = ctx.author
@@ -142,6 +144,7 @@ async def avatar(ctx, member: discord.Member = None):
 
 
 @client.command(aliases=["mc", "membercount"])
+@commands.cooldown(1, 4, commands.BucketType.user)
 async def members(ctx):
     member_count = len(ctx.guild.members)
     user_count = len([m for m in ctx.guild.members if not m.bot])
@@ -151,6 +154,7 @@ async def members(ctx):
 
 @client.command(pass_context=True, aliases=["clear", "purge"])
 @commands.has_permissions(administrator=True)
+@commands.cooldown(1, 4, commands.BucketType.user)
 async def clean(ctx, limit: int):
     await ctx.channel.purge(limit=limit + 1)
     await ctx.send(f'{ctx.author.mention} purged {limit} messages',
@@ -159,16 +163,25 @@ async def clean(ctx, limit: int):
 
 
 @client.command()
-@commands.has_permissions(ban_members = True)
-async def ban(ctx, member: discord.Member, *, reason = 'Unspecified'):
-    embed = discord.Embed(title=f'***Banned `{member}` for `{reason}`***', color=0x38805d)
-    await member.ban(reason = reason)
+@commands.has_permissions(ban_members=True)
+async def ban(ctx, member: discord.Member, *, reason='Unspecified'):
+    if member.id == ctx.author.id:
+        await ctx.send("You cannot ban yourself")
+        return
+
+    if member.top_role >= ctx.author.top_role:
+        await ctx.send(f"You can only ban members below your top role")
+        return
+
+    embed = discord.Embed(
+        title=f'***Banned `{member}` for `{reason}`***', color=0x38805d)
+    await member.ban(reason=reason)
     await ctx.send(embed=embed, delete_after=10)
     await ctx.message.delete()
 
 
 @client.command()
-@commands.has_permissions(administrator = True)
+@commands.has_permissions(administrator=True)
 async def unban(ctx, *, member):
     banned_users = await ctx.guild.bans()
     member_name, member_discriminator = member.split("#")
@@ -181,6 +194,63 @@ async def unban(ctx, *, member):
             await ctx.send(f'Unbanned `{member}`', delete_after=10)
             await ctx.message.delete()
             return
+
+
+@client.command(description="Mute command yay!!", aliases=["shut"])
+@commands.has_permissions(manage_messages=True)
+async def mute(ctx, member: discord.Member, *, reason='Unspecified'):
+    if member.id == ctx.author.id:
+        await ctx.send("You cannot mute yourself")
+        return
+
+    if member.top_role >= ctx.author.top_role:
+        await ctx.send(f"You can only mute members below your top role")
+        return
+    guild = ctx.guild
+    mutedRole = discord.utils.get(guild.roles, name="Muted")
+
+    if not mutedRole:
+        mutedRole = await guild.create_role(name="Muted")
+
+        for channel in guild.channels:
+            await channel.set_permissions(mutedRole, speak=False, send_messages=False, read_message_history=True, read_messages=False)
+
+    await member.add_roles(mutedRole, reason=reason)
+    embed = discord.Embed(title=f'Muted `{member}` for `{reason}`')
+    await ctx.send(embed=embed, delete_after=10)
+    await member.send(f"You were muted in `{guild.name}` for `{reason}`. DM a staff/owner if you feel this is unjustified")
+    await ctx.message.delete()
+
+
+@client.command(description="Unmute command yay!!")
+@commands.has_permissions(manage_messages=True)
+async def unmute(ctx, member: discord.Member):
+    mutedRole = discord.utils.get(ctx.guild.roles, name="Muted")
+
+    await member.remove_roles(mutedRole)
+    embed = discord.Embed(title=f'Unmuted `{member}`')
+    await ctx.send(embed=embed, delete_after=10)
+    await member.send(f"You were unmuted in `{ctx.guild.name}`")
+    await ctx.message.delete()
+
+
+@client.command(aliases=["abt", 'whois', 'who', 'ui', 'about'], description="Gets info about the user")
+async def userinfo(ctx, *, member:discord.Member = None):
+    if member == None:
+            member = ctx.message.author
+
+    embed=discord.Embed(
+      title=f'{member}'
+      )
+    embed.set_thumbnail(url=member.avatar_url)
+    embed.add_field(name="Name", value=f'`{member.name}`')
+    embed.add_field(name="Nickname", value=f'`{member.nick}`')
+    embed.add_field(name="ID", value=f'`{member.id}`')
+    embed.add_field(name="Account Created",value=f'<t:{int(member.created_at.timestamp())}>')
+    embed.add_field(name="Joined",value=f'<t:{int(member.joined_at.timestamp())}>')
+    members = sorted(ctx.guild.members, key=lambda m: m.joined_at)
+    embed.add_field(name="Join Position", value=f'`{str(members.index(member)+1)}`')
+    await ctx.send(embed=embed)
   
   
 client.run(os.environ['token'])
